refactor(processors): improve NLP pipeline, enrichment, and sentiment processing
- pipeline/enrichment.py: new single-pass enrichment orchestrator replacing
  scattered processor calls throughout the runner
- processors/sentiment.py: add batch processing, RoBERTa integration, and
  multi-window aggregation
- processors/keywords.py: cache YAKE extractor instances to avoid re-init
- processors/matching.py: fix title matching edge cases
- processors/sarcasm.py: harden model load error handling
- processors/emotions.py: harden model load and inference error handling
- processors/spam_detector.py: improve rule-based classification
- processors/metrics.py: remove obsolete standalone metrics computation
- utils/translation.py: add TranslationStatus enum and per-text status tracking

+++ src/sam/processors/matching.py
@@ -89,7 +89,7 @@ try:
     def _ratio(a: str, b: str) -> float:
         if not a or not b:
             return 0.0
-        return _rapidfuzz_ratio(a, b) / 100.0  # rapidfuzz returns 0-100
+        return float(_rapidfuzz_ratio(a, b) / 100.0)  # rapidfuzz returns 0-100
 
     _BACKEND = "rapidfuzz"
 except ImportError:
